requirements.txt
asgiref==3.11.0
crispy-bootstrap4==2025.6
Django==6.0
django-crispy-forms==2.5
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3
djangorestframework-3.16.1
